feat: support giving a regex pattern where the first match will be used as the body
Adds in a new feature: body_pattern. This regular expression will be matched against
the email body (before the body_template is applied) and the first match that is valid
will be used as the new email body. The $body variable that can be used by the body_template option will store the result of the match against body_pattern. If no matches are found, the program will terminate.